added a player file and tictactoe that calls the other files
