+++ tictactoe/__init__.py
@@ -1,2 +1,3 @@
-from tictactoe.tictactoe import TicTacToe
+from tictactoe.main_game import TicTacToe
 from tictactoe.board import Board
+from tictactoe.player import Player
